fix: styles naming
