+++ src/shared/components/FeedbackItem/index.module.scss
@@ -8,7 +8,7 @@
   font-family: var(--ffm-default);
 }
 
-.wrapper_info {
+.wrapperInfo {
   display: flex;
 }
